Add write-through cache in front of on-disk store

+++ server/src/main/java/com/proofpoint/discovery/store/PersistentStore.java
@@ -18,9 +18,7 @@ package com.proofpoint.discovery.store;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.smile.SmileFactory;
-import com.google.common.base.Function;
 import com.google.common.base.Throwables;
-import com.google.common.collect.Iterables;
 import com.proofpoint.log.Logger;
 import org.iq80.leveldb.DB;
 import org.iq80.leveldb.Options;
@@ -31,25 +29,42 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.Map;
 
-import static com.google.common.base.Predicates.notNull;
-
 public class PersistentStore
     implements LocalStore
 {
     private static final Logger log = Logger.get(PersistentStore.class);
     private final DB db;
     private final ObjectMapper mapper = new ObjectMapper(new SmileFactory()).disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
+    private final InMemoryStore cache;
 
     @Inject
-    public PersistentStore(PersistentStoreConfig config)
+    public PersistentStore(PersistentStoreConfig config, ConflictResolver resolver)
             throws IOException
     {
         db = Iq80DBFactory.factory.open(config.getLocation(), new Options().createIfMissing(true));
+        cache = new InMemoryStore(resolver);
+
+        for (Map.Entry<byte[], byte[]> dbEntry : db) {
+            try {
+                cache.put(mapper.readValue(dbEntry.getValue(), Entry.class));
+            }
+            catch (IOException e) {
+                byte[] key = dbEntry.getKey();
+                log.error(e, "Corrupt entry " + Arrays.toString(key));
+
+                // delete the corrupt entry... if another node has a non-corrupt version it will be replicated
+                db.delete(key);
+            }
+        }
     }
 
     @Override
-    public boolean put(Entry entry)
+    synchronized public boolean put(Entry entry)
     {
+        if (!cache.put(entry)) {
+            return false;
+        }
+
         byte[] dbEntry;
         try {
             dbEntry = mapper.writeValueAsBytes(entry);
@@ -59,49 +74,28 @@ public class PersistentStore
         }
 
         db.put(entry.getKey(), dbEntry);
-        return true; // todo implement check for already in db
+        return true;
     }
 
     @Override
     public Entry get(byte[] key)
     {
-        try {
-            return mapper.readValue(db.get(key), Entry.class);
-        }
-        catch (IOException e) {
-            throw Throwables.propagate(e);
-        }
+        return cache.get(key);
     }
 
     @Override
-    public boolean delete(byte[] key, long timestamp)
+    synchronized public boolean delete(byte[] key, long timestamp)
     {
+        if (!cache.delete(key, timestamp)) {
+            return false;
+        }
         db.delete(key);
-        return true; // todo implement check for no change
+        return true;
     }
 
     @Override
     public Iterable<Entry> getAll()
     {
-        return Iterables.filter(Iterables.transform(db, new Function<Map.Entry<byte[], byte[]>, Entry>()
-        {
-            @Override
-            public Entry apply(Map.Entry<byte[], byte[]> dbEntry)
-            {
-                try {
-                    return mapper.readValue(dbEntry.getValue(), Entry.class);
-                }
-                catch (IOException e) {
-                    byte[] key = dbEntry.getKey();
-                    log.error(e, "Corrupt entry " + Arrays.toString(key));
-
-                    // delete the corrupt entry... if another node has a non-corrupt version it will be replicated
-                    db.delete(key);
-
-                    // null if filtered below
-                    return null;
-                }
-            }
-        }), notNull());
+        return cache.getAll();
     }
 }
